chore: refactoring and some checks
