+++ api/src/main/kotlin/org/jetbrains/kotlinx/dl/api/core/layer/reshaping/Permute.kt
@@ -1,8 +1,9 @@
-package org.jetbrains.kotlinx.dl.api.core.layer.core
+package org.jetbrains.kotlinx.dl.api.core.layer.reshaping
 
 import org.jetbrains.kotlinx.dl.api.core.KGraph
 import org.jetbrains.kotlinx.dl.api.core.layer.Layer
-import org.jetbrains.kotlinx.dl.api.core.shape.TensorShape
+import org.jetbrains.kotlinx.dl.api.core.shape.shapeFromDims
+import org.jetbrains.kotlinx.dl.api.core.shape.toLongArray
 import org.tensorflow.Operand
 import org.tensorflow.Shape
 import org.tensorflow.op.Ops
@@ -19,15 +20,22 @@ public class Permute(
     public val dims: IntArray,
     name: String
 ) : Layer(name) {
+
+    init {
+        require(dims.sorted() == (1..dims.size).toList()) {
+            "The values in dims should be consecutive and start from 1."
+        }
+    }
+
     override fun build(tf: Ops, kGraph: KGraph, inputShape: Shape) {}
 
     override fun computeOutputShape(inputShape: Shape): Shape {
-        val outputShape = LongArray(dims.size + 1) { 0 }
+        val outputShape = inputShape.toLongArray()
         dims.forEachIndexed { i, dim ->
-            val target_dim = inputShape.size(dim);
+            val target_dim = inputShape.size(dim)
             outputShape[i + 1] = target_dim
         }
-        return TensorShape(outputShape).toShape()
+        return shapeFromDims(*outputShape)
     }
 
     override fun forward(
@@ -38,7 +46,7 @@ public class Permute(
     ): Operand<Float> {
         val permArray = intArrayOf(0) + dims
         val perm = tf.constant(permArray)
-        return tf.linalg.transpose(input, perm);
+        return tf.linalg.transpose(input, perm)
     }
 
     override var weights: Map<String, Array<*>>
